[ffx] Add always-on behaviors to femu
Some of the flags passed to the Femu emulator from device-launcher were
coded to be always on. This CL adds a catch-all Behavior for those
flags, so they will be included during startup regardless of other
configuration choices. These will likely be moved to independent
Behaviors as we proceed.

Change-Id: I4923851028f64e21ce79d10302bfcf85358189b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614990
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>