[driver_runtime] fdf_channel_call implementation.
This adapts the zircon channel call implementation and tests.

When a client calls Channel::Call, we allocate a new txid,
and write it to the message and install a new message waiter.

When the channel receives a message of a matching txid,
it will deliver the message to the waiter and signal it.

Bug: 87161

Change-Id: I9fe447856fe8ac77699daf7ebdf3b36257a3e13f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612140
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>