[run-test-suite] stdout reporter part 1: Shell writer
This is the first in a series of changes that moves the stdout output
produced by run-test-suite into a reporter implementation. We
previously added reporters to support structured output, but stdout is
currently output in the same logic as execution logic. This series of
changes is intended to remove stdout from execution logic so that we
don't need to carry around extra state only needed for stdout.

This change adds a Write implementation that serves as a 'handle' to an
underlying Write implementation. Multiple handles may be created for a
single underlying Writer. In this case output from each handle is
delimited by line breaks and merged into the underlying writer.
This will be used by the stdout reporter to give out multiple handles
that all write to stdout.

Change-Id: I5b43e0e1a4d32fac2be7c0e7c67a415944d02e1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601525
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>