[graphics][compute][spinel2] Initial drop of Spinel2
The CL includes Spinel2.

Spinel2 is a new version of Spinel that leverages timeline semaphores
and buffer references in order to present a "feed-forward" pipeline
that does not rely on host-pumped callbacks.

Spinel2 relies on the indirectly dispatched version of
`compute/radix_sort`.

Spinel2 prepares the way for new features:
 * Batch loading of path geometries
 * Batch loading of rasterized paths
 * Incrementally maintained compositions and stylings
 * Partial update of storage swapchains and Vulkan images.
 * Pixel and subpixel translation of placed rasters.
 * Simpler "feed-forward" interaction with Vulkan swapchains

Supported targets:

 * NVIDIA sm_35+

The previous list of targets will be added in a following CL.

Testability:

 * Build: fx set core.chromebook-x64 --with src/graphics/lib/compute/spinel2/platforms/vk/tests
 * GTest: out/core.chromebook-x64/host_x64/exe.unstripped/spinel-vk-test
 * Interactive: out/core.chromebook-x64/host_x64/exe.unstripped/spinel-vk-bench -f <svg>

Note that the interactive test is non-functional.

Change-Id: I59759985f5ac5e6d32c3569e0131d09c314c43c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615761
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>