[zircon][scripts] Fix doc generation for missing syscalls
When writing out stubs for a new .md file for missing syscalls, open the
file as a text file. Previously the file was being opened in binary
mode, but then a string was being written to it.

Change-Id: Iaa57fd8514b51266f9eeb65b021f61fae31ac20f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614985
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>