[input] Removes the now-unnecessary modifier tracking
ImeHandler used to track the modifier and lock states.  When we added
ModifierHandler, this became unnecessary, so long as we know that
ModifierHandler is installed ahead of ImeHandler. Since we should always
do that we should be justified to remove the redundant state tracking.

This change removes the redundant state tracking and adapts the tests
accordingly.  While this is mainly a refactoring change, some tests did
not uphold the modifiers invariant on the input events, so that had to
be modified.

The invariant in question is that at least one modifier bit must be set
on a key event that reports a Shift pressed event.

Fixed: 89763
Change-Id: Idd47ab5024ac710b5f94ec2f96ecb8559cf5788d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613782
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>