[component][fuzzing] Add SyncWait for diagnostics
The component fuzzing framework is still exhibiting some hard to trace
flakes that manifest as unexpected hangs in the RunnerImplTests.
Unfortunately, attempts to reproduce these flakes locally have been
unsuccessful. In order to isolate the root cause, this CL introduces a
wrapper for libsync that logs a warning if an indefinite wait exceeds a
configured threshold.

Bug: 86343
Change-Id: I133bec6ce7a2e6a26a5db4dc17171d5bb7c0ef35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616624
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>