[blobfs] Fix casting.
Enables the warning about implicit casting and fixes the affected code.

Converts most numeric casts to safemath::checked_cast. Test and metric
code is unchanged.

The blobfs host code in particular is confused about whether inodes and
block indices are 32 or 64 bits. This isn't fixed in this patch but I
tried to fix a couple of places.

Bug: 89819
Change-Id: I247b8b534de1345874f3405a956076f46962a5e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616881
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>