[wlan][policy] add metric to count roam scans
Log to cobalt when policy decides to kick off a scan to look for
networks to roam to.

Note about testing: the network scoring logic will be added in a
different CL and as of this change policy will never decide to perform a
roaming scan. So this change was manually tested with a simple scoring
function change and the unit test that requires a scoring function is in
a different CL.

Test: manually tested with local_log_changes and with a connection
      quality scoring function based on RSSI that the metric would
      be logged when a scan would be triggered.
      Added a unit test for telemetry, and a unit test in iface_manager
      is added in a future CL
Change-Id: I46fff1e226988522b06b6cc6dfe8d9ce4e43520e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606668
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>