[run-test-suite] stdout reporter part 4: update run-test-suite interface
This is part 4 in a refactor to move stdout output to a reporter
implementation. This makes a small update to the run-test-suite
interface so that callers are responsible for composing reporters and
passing them into run-test-suite.

Testing: existing integration tests

Change-Id: Iecbe1feb4843a5e86102d7f931ba7507a43572b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613632
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>