[component][fuzzing] Add Corpus::Load
This CL adds a way to load corpora from package storage. This allows
fuzzer authors to include their seed corpora in the fuzzing package and
supply the package-relative location as arguments to the target adapter.

The actual passing of those parameters from the target adapter to the
engine will be handled in the next CL.

Bug: 24740
Change-Id: Ief213cb927c5e2ada0670b8b3cdea7eeca9de132
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609060
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>