[sl4f] Create SingleWebDriverConnector
Unlike the existing WebDriverConnector, this new connector only connects
to one WebDriver session at a time, and hence require only one port to
be forwarded, which is useful in scenarios like QEMU with user-mode
networking. The new connector also has the added benefit of accepting a
custom HttpClient, which makes it more versatile in different networking
environments (e.g. proxy, HTTPS).

This requires some minor changes to TcpProxyController to allow reusing
a dropped proxy port. Some fine-level logging is added to
TcpProxyController to bring it on par with the SSH forwarder.

All changes are supposed to be backward-compatible.

In addition to unit tests, this change has been tested end-to-end on
internal infra.

Bug: b/180759363
Change-Id: I62fe4b077942894c9dd7d3b66e9caf7229c51a96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613021
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Robert Ma <robertma@google.com>