[fdf] Update driver references when we unbind
When a driver component is unbound, we need to update the node it is
bound to, so that the node will no longer wait for the driver to be
unbound during its own teardown. Without that, there is no way to
introspect the ServerBindingRef to be aware of the fact that the
AsyncBinding was torn down.

Multiply: interop_test
Fixed: 89937
Bug: 33183
Change-Id: Ieb143f32e397046204be8e9e33dbe3ac8004ecb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615466
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>