[sdk] Build exported SDK by default using symlinks
This will cause `sdk` type build targets to build their _export targets
and modifies the export script to use symlinks instead of copies, which
reduces the size of the exported core SDK from 1.2 GB to 3.3 MB. The
build takes less than 0.1s longer with `fx set core.x64`.

The objective here is to allow tools, primaily ffx, to interact with the
SDK in-tree in the same way as in out-of-tree SDK installations.

Bug: b/209679218
Change-Id: Ied3005734ff7592825819cdeee07a73b23c0b216
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616984
Reviewed-by: Misha Gridnev <gridman@google.com>
Commit-Queue: Sean Cuff <seancuff@google.com>