[fxl] Shuffle googletest cases
This causes googletest to emit the value of the random seed on each
iteration, which is nice considering we have several tests that
currently use this (and are probably all getting a random seed of zero).

Standardize retrieval of this value while I'm here.

Change-Id: Ie2466fb2092b04bcf9f7e6ddc5d4fecec25ff49f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616181
Commit-Queue: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Shai Barack <shayba@google.com>