[zircon][vm][ppb] Replace with loaned on move to active
We replace non-loaned pages with loaned pages as we put
a page in it's designated queue, if the designated queue
is an active queue.  Under normal conditions this
gradually replaces non-loaned pages with loaned pages.
Near OOM, as we're peeking pretty much all the pages
except active ones, we'll replace any pages we're moving
into the active queues.

This doesn't quite completely guarantee that we'll be
using all available loaned pages by the time we OOM, but
this gets us a lot closer to that.  To get the rest of
the way we could do a final sweep of the active queues
after eviction near OOM is complete but before we
consider eviction near OOM done.

Change-Id: I458fe6fc9921a9f7ce1d3075596ac500dff35042
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614224
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>