[testrunner] Refactor ffx testing logic into an FFXTester.
Before, we included the ffxInstance as part of the FuchsiaSSHTester and
would use it if ffx was enabled. Now we separate it into its own Tester
and include an SSHTester as part of the FFXTester to run v1 tests and
fetch data sinks (not supported by ffx yet). Once all v1 tests have
migrated to v2 and fetching data sinks is supported by ffx, we can
remove the SSHTester from the FFXTester. Since there will be more logic
that needs to be added to handle processing test outputs from ffx, it
makes more sense to create a separate FFXTester rather than add it to
the FuchsiaSSHTester.

Bug: 89229
Change-Id: I84c3bd3d68b2d1e57ae6c4f0c389ed495d52970f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616632
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>