[realm_builder] add GetRealmDecl/ReplaceRealmDecl
The GetComponentDecl and ReplaceComponentDecl functions are only capable
of interacting with children of a realm. Interacting with the decl of
the realm itself is also an important use-case, so that users may do
things like add environments or declare collections in a realm.

Add GetRealmDecl and ReplaceRealmDecl to the realm builder FIDL to
address these use-cases.

Change-Id: I13d6e63debb2c8491432bcec84c31d95c076b080
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617141
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>