[component][fuzzing] Add TargetAdapterClient
This CL adds a client to fuchsia.fuzzer.TargetAdapter. Both the engine
and test-engine for the framework need to interact with the target
adapter in same manner, so it is useful to extract these interactions
from RunnerImpl and encapsulate them in their own class.

Bug: 24740
Change-Id: If8151beec76afb70d397c12d42c727e01a036288
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603974
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>