[time] Move tests to absolute time where beneficial.
In a couple of cases the fact that a test couldn't directly control when
a clock update occured led to a less accurate test. This fixes those
without changing the logic of how time is maintained.

Note than in many cases a test just needed to start a clock or set a
coarse time, and in those cases approximate_time is fine.

Bug: 65726
Change-Id: I9a5ea4068a13273494949787728b6069470d0d5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617782
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>