[lib][concurrent] Add libconcurrent to the kernel. (RE-LAND)
This CL is a re-land of http://fxr/606085 which was reverted in
http://fxr/616782 because of http://fxb/90171

Please refer to the bug for details.  The original commit message was
as follows.

"""
Add libconcurrent to the kernel as kconcurrent, defining lockdep
guards in the process.  Also add unittests for the guards.
"""

Fixed: 90171
Change-Id: If7b46a160d1f2b0c6bfcf6fa37278c5cec6c8d33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/618782
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>