[password auth]Return subdirectory from GetDataDirectory.
Sorry I forgot to comment on this with the original CL but really we
should be returning a subdirectory of the account partition from
password authenticator rather than the root. That way we have the option
of creating top level files for the account system (e.g. sealed account
metadata) that clients could never read, or of one day giving
different subdirectories to different clients.

Change-Id: Iec69a50269f4118cb13c530a16a6677db2e6d9d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610133
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>