[wlan][policy] Add connection duration roam metric
Log a metric for how long the device was connected to an AP before
roaming to another AP or how long it was connected before an unexpected
disconnect and the total of both to compare.

Currently the decision of whether or not a roam should happen is not
implemented, but when it is implemented it will send a telemetry event
for the proactive network switch.

Test: added unit tests: fx test wlancfg-test,
      manually tested with also_log_locally for non-roam disconnectss
Change-Id: I39693a929e30b3592a351812ed8476425c41d113
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616661
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>