[debugger] Handle stripped ELF symbols.
Stripped ELF binaries are missing the .symtab section. In this case the
only ELF symbols we can use are in the .dynsym section.

This merges the symbols from both .symtab and .dynsym when reading ELF
symbols so we don't miss any, regardless of the stripped status of the
binary.

Change-Id: Ifef4e843ebfddff912ab43dea83da3cbc10ede68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619081
Reviewed-by: Dangyi Liu <dangyi@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>