[kernel][hypervisor][arm64] Do not convert paddr
The pmm arena base is already a paddr_t and the previous code that
treated it as a vaddr and attempted to convert it to a paddr was
incorrect.

This bug was benign for so long since the relevant ARM boards
that virtualization runs on have their physical memory starting at 0,
and all of this memory would be presented to the kernel as one large
arena. Although the conversion of vaddr 0 to a paddr would fail, it
would fail by giving a value of 0, which coincidentally was the right
value.

Change-Id: I951791ecaf58c251d32aaf9e426a3069d11905cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619302
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>