[zircon][vm][ppb] vmo_contiguous_decommit_test fixup
Have vmo_contiguous_decommit_test() cover the case where
a page is definitely borrowed, then reclaimed by the
loaning contiguous VMO before the borrowing VMO is deleted.

Previously this case was broken for a time (double-free
due to bad merge), but not failing CQ.  Thanks to kostyak@
for noticing this and pointing it out.

Confirmed that this test now fails if I put back the
double-free locally.

BUG: 83252
Change-Id: Ie748ffbbf119879c7b4e866924c2b64cc8aafe59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619330
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>