[netstack3] Check IdMap contents in property test
The existing property test checks that arbitrary sequences of operations
don't leave the contents of the IdMap in an invalid state. This patch
extends that by adding a HashMap that represents "a source of truth". By
applying the same operations to both the IdMap and HashMap, we can
verify that the IdMap implements its API faithfully. As a side-effect,
this cleans up the `let _ = map.Xxx()` calls in Operation::apply since
the results of applying operations on the map are now used instead of
ignored.

Change-Id: I50f0cd2b654103978c72ac9e9e6c357f3c773090
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615661
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>