[test][kernel_panic] Add SMEP/PXN tests.
This CL adds an emulator test that checks that the kernel crashes when
it tries to execute user memory.

A new k command is added: `k crash_user_execute` creates an
userspace mapping, copies a simple function into it, remaps it as
read-only executable and calls that code.

The test should cause a kernel crash in both arm64 (PXN) and x86-64 (SMEP).

This CL also renames the SMAP test so it is architecture agnostic, to
avoid confusions.

Test: fx test kernel_panic_tests
Bug: 88451
Change-Id: Ib5ba6680bbfa151a06ab2353280c8896459f957a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611946
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>