[timekeeper] Don't automatically start monitor clock.
We're not currently using a monitor clock, but the code to create it
currently follows the old primary clock behavior where component
manager used to start the clock before giving it to us. Its more
consistent to create a clock that hasn't started and let the
clock_manager take care of the start when its ready.

Bug: 65726
Change-Id: I55fe5bd7ac06cb500ae208b2af1830a831343f6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617710
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>