[fit] Replace bind_member with more-efficient variant
This CL deprecates the existing version of fit::bind_member,
which uses the fit::bind_member(&instance, &Obj::Method) syntax and
introduces a new variant with a different API
fit::bind_member<&Obj::Method>(&instance).

The motivation for this is that the existing fit::bind_member is
inefficient and the lambda is 24-bytes, leading to a heap allocation
in a fit::function. In contrast, the new fit::bind_member's lambda
only requires 8-bytes.

The idea is to migrate existing usages to the new fit::bind_member
where possible (C++17 and up) and migrate to a regular lambda in
other cases (C++14 and down).

Change-Id: Ifac9b7434cf417ca59471d29aa97884113d16b7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612883
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>