[llcpp_drivers] Generate skeleton driver code
This CL adds *_driver_wire_messaging.{h,cc} files and generates output
in those files for protocols marked with @transport("Driver"). For now,
the output is specialized for channels, but over time it will be updated
to be specialized for drivers instead.

In addition, this CL adds golden file support and introduces two goldens
- one for one way calls and one for two way calls. The one way call test
also covers domain objects so it is possible to ensure that domain
objects are reused correctly.

Change-Id: I5fa3cf748b564f54ba42dc98dc444ce1ea2caeca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617264
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>