[netcfg] Migrate to diagnostics_log and tracing
Currently the logs emitted by netcfg are tagged, by the default behavior
in the log crate, with the full module path of where they originated.
This means that a log from the `devices` module gets tagged as
'netcfg::devices'. We want all logs from the netcfg binary to get the
same tag so that e.g. filters apply equally to all netcfg logs.

Use diagnostics_log and tracing instead of fuchsia_syslog and log in
order to get this behavior.

Bug: 90277
Change-Id: I3dc9a63a3461f4bcb95b7c87ec73d5e14d8c4f5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619824
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>