[dhcpv6] Extract address assignment test setup fn
Introduce a test setup fn to create a client in stateful mode, and
perform message exchange to assign the configured addresses to the
client. The functionality of the test setup fn is extracted from the
`assign_addresses` test, with the addition of making the addresses to be
assigned configurable. The test fn is intended to be used as test setup
for future state machine testing starting in `AddressAssigned` state.

Add an utilitary Identity Associsation test type to configure the
addresses to be assigned, specifying T1/T2 values; intended to be used
in future Renew/Rebind testing as well.

Bug: 88838

Change-Id: I955e429967f1a7eaeb710ee2a941415c2c5eec99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604122
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>