[component_manager] Make `abs_moniker` a private field in `ComponentInstance`
This CL is one of several to accomplish the goal of hiding
`AbsoluteMoniker` from public APIs. This change makes private the
`abs_moniker` field and introduces the `partial_abs_moniker` field to
`ComponentInstance` and `ComponentInstanceForAnalyzer`.

The following changes are made:

- Occurrences of `<component>.abs_moniker.to_partial()` are replaced
  with a direct reference to `<component>.partial_abs_moniker`.
- All other occurrences of `<component>.abs_moniker` are replaced with
  a call to the `ComponentInstanceInterface` trait fn `<component>.abs_moniker()`.
- Logging/debugging messages that displayed `<component>.abs_moniker` now
  displays `<component>.partial_abs_moniker`.

The reason for this change set is to reduce and coalesce external
references to `abs_moniker` so that a future CL can remove or replace
public references entirely (including the `abs_moniker()` trait fn).

Bug: 87407
Change-Id: Ia3ab346116c4619ad78313a12c4756722ddb239e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619001
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>