[netstack3] Split IP device state by version
Store IP device state in an `IpDeviceState` struct which is generic on
IP version.

Migrate some functionality to be methods on that type. Move that type
along with some other state types to a separate `state` sub-module so
that other modules under `device` cannot access their internals, and
invariants are easier to maintain.

Add a `get_ip_device_state` method to `NdpContext`. Remove some methods
which are redundant given that method, and give some methods default
impls in terms of that method.

Bug: 86168
Bug: 48578
Change-Id: I8dd8a39158644fc2883bdd423f38d021f1298f68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582921
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>