[weave] Mark success on duplicate removals
The WARM state machine stops if we fail to remove an address when
requested, but this can happen in situations where the interface was
already removed in between calls to WARM. In these cases, produce a
warning log but allow the state machine to progress. Tests are added to
confirm behaviors on duplicate add and remove.

Bug: 89993
Test: fx test weavestack-adaptation-unittests
Change-Id: I53714ce9e4cb409b1591b654d80b01cfa10e8e3f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620006
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Fuchsia-Auto-Submit: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>