[kernel][vm] Add dirty page queue
Add PageQueuePagerBackedDirty which holds pager-backed pages that are
not clean, i.e. any pages with dirty_state AwaitingClean or Dirty. These
pages cannot be evicted and should be moved out of the normal
pager-backed queues which are used to find eviction candidates.

Pages are moved to the dirty queue on transition from Clean to Dirty,
and moved out on transition from AwaitingClean to Clean.

Test: kernel-unittests

Bug: 63989
Change-Id: I31f8e23db22c7471ec0b172fa5e10f33ac5255e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614121
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>