[blobfs][test] Speed up use-second-bitmap-test
This test requires generating a blob that is block_size^2 (!!!) bytes
large.  The obvious way to speed it up would be to reduce block_size,
but that isn't feasible, given how deeply baked-in the assumption that
blobfs' block size is 8192 is.

Make a few tactical changes to hopefully speed the test up:
- Disable compression, which is wasted effort
- Reduce the file size from ~300M to ~256M
- Make the "generator" function a NOP so the blob data is just
  uninitialized garbage

Locally this results in a ~2x improvement.

Bug: 89815
Bug: 90384
Change-Id: I43a982e0c60bfccbdbd0c65a039b51f662858a3f
Multiply: blobfs-use-second-bitmap
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619921
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>