[ffx] Remove previous Behavior code from ffx emu
We're going a different direction with the idea of Behaviors in ffx emu.
We'll be utilizing HandleBars to simplify the specification and code to
handle it.

This CL prepares for the conversion by removing all of the previous
iteration of Behaviors from the plugin's code, as well as the PBM code.

Change-Id: Icb939492f492f5b4bdabe3e148f7240fad66f844
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620102
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>