[build] Fix variant selection in fuzzer_package() template.
A small CL to fix the way variant selection is performed in the
fuzzer_package() template. The old code could only work with
very restricted selectors definitions, and would not work
with more complex ones, for example:

  {
    variant = "asan-fuzzer"
    output_name = [
      "first_fuzzer",
      "second_fuzzer",
    ]
  }

Because `output_name` is a list with two items here.
Or even:

  {
    variant = "ubsan-fuzzer"
    dir = [ "//src/foo" ]
  }

Because the `dir` was not supported.

+ Update a comment in //build/zircon/BUILD.gn to point to
  the right .gni file.

Bug: NONE
Change-Id: I52924b9a6409fdb7ee330c0bcda00dfcbc40a1e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619644
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>