[fx] Make manifest mismatch error actionable
This most often happens when a developer with a fully internal checkout
tries to reproduce an internal build that used the internal integration
repo, but with a public manifest.

In that case the public manifest contains a subset of the internal
manifest that the developer is likely using locally, so syncing to the
revision used by the build is probably good enough.

However, this is not sufficient for CQ builds (where "patches" is
non-empty) because the manifests need to match in order for jiri to
patch in the change used by the build.

Change-Id: I489919794fd1d28149fab8a7e8ed05e180effdb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620781
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>