[kernel][vm] Fit lambda within kDefaultInlineCallableSize
Reduce the capture size of a lambda used in
PagerDispatcher::QueryDirtyRanges to fit within
kDefaultInlineCallableSize which is 16 bytes. Previously the lambda was
greater than 16 bytes but was getting the correct behavior when used to
initialize an inline fbl::Function sized 32 bytes due to alignment
(16 bytes for the capture, 8 for the function pointer and the remaining
for alignment). Do not rely on alignment and instead size the capture
within kDefaultInlineCallableSize as expected.

Bug: 63989
Change-Id: I7bbc9c07afa8fd69fecdac5439ff142c79339325
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620889
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>