[ffx] Add fmt::Display to virtual device enums
We found ourselves copying the same code repeatedly for various
enumeration definitions throughout the virtual device/emulator engine
code, enabling the enums to be cast to and from strings. This is because
enums cannot derive those properties automatically.

This CL adds a public macro to the SDK metadata type definition library,
which derives the fmt::Display and FromStr traits on those enums. It
also uses the same macro on the ffx emu plugin enums. This reduces the
amount of code in both of these locations, while simplifying any future
additions or changes to those types.

Bug: fxbug.dev/90481

Change-Id: I3221808d32d33e2b79093c1e224a7129b079b5e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620103
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>