[ffx] Move command line flags to emulator config
This CL adds a FlagData type and field to the emulator configuration.
This is similar to the previous Behaviors map, except it's all in one
structure instead of a map of them. There are still the same four
fields: args, features, kernel_args, and options. These map to the parts
of the command line they represent.

We also remove extraneous fields from the engines, and adjust the
build_emu_cmd accordingly.

Bug: fxbug.dev/90481

Change-Id: I5b936e734963fce239ab1da2cda787ea494f2e29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620104
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>