[wlan] Create a PriorityQueue driver component
Create a PriorityQueue intended for use with priorities based on 802.1q
priority code points. This is intended for drivers that manage queues
internally in the driver.

Because it's based in 802.1q priority code points we can limit the
number of different priorities and therefore allow O(1) insertion
instead of a more generic priority queue which might need O(log n).

Additionally the queue allows popping of frames with a limited set of
priorities such that a user can choose to only pop highly prioritized
frames in situations where resources might be constrained.

Bug: 79622
Test: Included unit test
Change-Id: I23da8c251702363fece96d6f3a380c984f7cb3ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/562266
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: John Sheu <sheu@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>