[run-test-suite] Add back test name preamble
testparser relies on this preamble to exist to recognize the test as
a test run using test framework. Removing this broke testparser and
it could no longer collect information on test cases.
While we're moving towards a world in which tools have a structured
output to parse, the migration isn't complete and we shouldn't break
tools in the meantime.

Change-Id: I03247d684fe495ed70dffeb899b278fcca64cbfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621145
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>