[virtualization] Use an FVM partition for the stateful image
Update the linux_runner so that when we use a stateful image, we will
look for an FVM partition to back the virtio-block device that is used
for the stateful image. If no such FVM partition exits, than we will
create it.

Bug: 90276
Bug: 90469
Change-Id: I27c1fb96a08f349c533cd6698d1ee00458658642
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614422
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>