[virtualization][linux_runner] Add LinuxManager FIDL.
This exposes a LinuxManager protocol from the linux runner
that can be used to launch a linux guest and monitor the
startup process.

Vsh has been updated to take advantage of this and the
vsh-terminal component can as a result be used to launch
and enter a Linux container.

Test: vsh -c
Change-Id: Ie4d6f7de08c9e3dab3b90968a591b84e9c02c70b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620981
Reviewed-by: Abdulla Kamar <abdulla@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Reveman <reveman@google.com>
Reviewed-by: Jonathan Bailey <jonathanbailey@google.com>