[assembly] Fix assembly inputs collection
This change does two things
1. includes the package manifests as inputs
2. drop all the inputs into "built/artifacts/"

We are putting all the artifacts into "built/artifacts/" so that when a
path has the "../../" prefix it will end up in the root. This makes
OOT integration much easier, because paths in the product config, board
config, and package manifests do not need to be modified.

Bug: 85370
Change-Id: Ic172abe2f3059cea0859e851b39ae35d859fe7ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621664
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>