[diagnostics][sampler] Better keys for cached values
To send event-count diffs to Cobalt we need to remember previous
values. Currently these are keyed by the Diagnostics hierarchy path that
selected them. But it can happen that multiple Diagnostics files (payloads)
are returned for a single component, and a selector can match a value from
each.

To avoid name collisions, we'll build the metric_cache key from the
metadata.filename as well as the selector string.

This CL refactors ProjectSampler to make it more unit-testable.

Fixed: 90216
Test: $ fx test //src/diagnostics/sampler
Change-Id: I60d365533c73c02640f989eaaa9b532e3a0b5563
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617705
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Chris Phoenix <cphoenix@google.com>