[ffx][test] Save the ffx stdout as a 'report' artifact
This adds a new artifact type to the structured output, a 'report'
artifact that contains the stdout generated by ffx. This artifact is
useful while we transition testrunner over to using the structured
format, as it allows us to do so without affecting the output a
developer sees in CI.

Implementation is done by creating a new reporter which composes a
directory reporter and shell reporter.

Testing: new unit tests, manually verify the report is present after
running 'ffx test'.

Change-Id: I8260a4e5544160cbea6c8f53acef1daf8ea13c78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616362
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>