[ui-input][input-pipeline] add mouse_binding::MousePhase
Add the moue_binding::MousePhase enum, which represents the phases
which are relevant to a mouse device.

This eliminates the need for upper layers (e.g. ClickDragHandler)
to validate that the phase of a MouseEvent is valid for a mouse.

Bug: 90516
Test: unit tests
Change-Id: I2dff2b236adcfa9a55dbc7eb90789d5965dd42e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621281
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>